pyproject.toml
[project]
name = "health-connect"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "bcrypt>=4.3.0",
    "email-validator>=2.1.0",
    "fastapi>=0.104.0",
    "jinja2>=3.1.2",
    "passlib>=1.7.4",
    "pydantic>=2.4.2",
    "pydantic-settings>=2.0.3",
    "python-dotenv>=1.0.0",
    "python-jose>=3.3.0",
    "python-multipart>=0.0.6",
    "sqlalchemy>=2.0.23",
    "uvicorn>=0.24.0",
]
